+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes qsf-integration project will be documented in this file.
 ## [UNRELEASED]
 ### Added
 - add a pausedUntil to the monitoring builder
+- implement a facet filter
 
 ### Changed